Selection : handle frame selection

+++ src/valeditor/editor/Selection.hx
@@ -1,5 +1,7 @@
 package valeditor.editor;
 import openfl.events.EventDispatcher;
+import valedit.ValEditFrame;
+import valeditor.ValEditorFrameGroup;
 import valeditor.events.SelectionEvent;
 
 /**
@@ -11,22 +13,30 @@ class Selection extends EventDispatcher
 	public var object(get, set):Dynamic;
 	private function get_object():Dynamic 
 	{
-		if (this._objectGroup.numObjects == 0 && this._templateGroup.numTemplates == 0) return null;
+		if (this._frameGroup.numFrames == 0 && this._objectGroup.numObjects == 0 && this._templateGroup.numTemplates == 0) return null;
+		if (this._frameGroup.numFrames == 1) return this._frameGroup.getFrameAt(0);
 		if (this._objectGroup.numObjects == 1) return this._objectGroup.getObjectAt(0);
 		if (this._templateGroup.numTemplates == 1) return this._templateGroup.getTemplateAt(0);
+		if (this._frameGroup.numFrames != 0) return this._frameGroup;
 		if (this._objectGroup.numObjects != 0) return this._objectGroup;
 		return this._templateGroup;
 	}
 	private function set_object(value:Dynamic):Dynamic
 	{
-		if (value == null && this._objectGroup.numObjects == 0 && this._templateGroup.numTemplates == 0) return value;
+		if (value == null && this._frameGroup.numFrames == 0 && this._objectGroup.numObjects == 0 && this._templateGroup.numTemplates == 0) return value;
+		if (this._frameGroup.numFrames == 1 && this._frameGroup.getFrameAt(0) == value) return value;
 		if (this._objectGroup.numObjects == 1 && this._objectGroup.getObjectAt(0) == value) return value;
 		if (this._templateGroup.numTemplates == 1 && this._templateGroup.getTemplateAt(0) == value) return value;
+		this._frameGroup.clear();
 		this._objectGroup.clear();
 		this._templateGroup.clear();
 		if (value != null)
 		{
-			if (Std.isOfType(value, ValEditorObject))
+			if (Std.isOfType(value, ValEditFrame))
+			{
+				this._frameGroup.addFrame(cast value);
+			}
+			else if (Std.isOfType(value, ValEditorObject))
 			{
 				this._objectGroup.addObject(cast value);
 			}
@@ -42,10 +52,12 @@ class Selection extends EventDispatcher
 	public var numObjects(get, never):Int;
 	private function get_numObjects():Int
 	{
+		if (this._frameGroup.numFrames != 0) return this._frameGroup.numFrames;
 		if (this._templateGroup.numTemplates != 0) return this._templateGroup.numTemplates;
 		return this._objectGroup.numObjects;
 	}
 	
+	private var _frameGroup:ValEditorFrameGroup = new ValEditorFrameGroup();
 	private var _objectGroup:ValEditorObjectGroup = new ValEditorObjectGroup();
 	private var _templateGroup:ValEditorTemplateGroup = new ValEditorTemplateGroup();
 
@@ -54,9 +66,31 @@ class Selection extends EventDispatcher
 		super();
 	}
 	
+	public function addFrame(frame:ValEditFrame):Void
+	{
+		if (frame == null) return;
+		this._objectGroup.clear();
+		this._templateGroup.clear();
+		this._frameGroup.addFrame(frame);
+		SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
+	}
+	
+	public function addFrames(frames:Array<ValEditFrame>):Void
+	{
+		if (frames == null || frames.length == 0) return;
+		this._objectGroup.clear();
+		this._templateGroup.clear();
+		for (frame in frames)
+		{
+			this._frameGroup.addFrame(frame);
+		}
+		SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
+	}
+	
 	public function addObject(object:ValEditorObject):Void
 	{
 		if (object == null) return;
+		this._frameGroup.clear();
 		this._templateGroup.clear();
 		this._objectGroup.addObject(object);
 		SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
@@ -65,6 +99,7 @@ class Selection extends EventDispatcher
 	public function addObjects(objects:Array<ValEditorObject>):Void
 	{
 		if (objects == null || objects.length == 0) return;
+		this._frameGroup.clear();
 		this._templateGroup.clear();
 		for (object in objects)
 		{
@@ -76,6 +111,7 @@ class Selection extends EventDispatcher
 	public function addTemplate(template:ValEditorTemplate):Void
 	{
 		if (template == null) return;
+		this._frameGroup.clear();
 		this._objectGroup.clear();
 		this._templateGroup.addTemplate(template);
 		SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
@@ -84,12 +120,18 @@ class Selection extends EventDispatcher
 	public function addTemplates(templates:Array<ValEditorTemplate>):Void
 	{
 		if (templates == null || templates.length == 0) return;
+		this._frameGroup.clear();
 		this._objectGroup.clear();
 		for (template in templates)
 		{
 			this._templateGroup.addTemplate(template);
 		}
 		SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
+	}
+	
+	public function hasFrame(frame:ValEditFrame):Bool
+	{
+		return this._frameGroup.hasFrame(frame);
 	}
 	
 	public function hasObject(object:ValEditorObject):Bool
@@ -100,6 +142,31 @@ class Selection extends EventDispatcher
 	public function hasTemplate(template:ValEditorTemplate):Bool
 	{
 		return this._templateGroup.hasTemplate(template);
+	}
+	
+	public function removeFrame(frame:ValEditFrame):Void
+	{
+		var removed:Bool = this._frameGroup.removeFrame(frame);
+		if (removed)
+		{
+			SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
+		}
+	}
+	
+	public function removeFrames(frames:Array<ValEditFrame>):Void
+	{
+		var frameRemoved:Bool = false;
+		var removed:Bool;
+		for (frame in frames)
+		{
+			removed = this._frameGroup.removeFrame(frame);
+			if (removed) frameRemoved = true;
+		}
+		
+		if (frameRemoved)
+		{
+			SelectionEvent.dispatch(this, SelectionEvent.CHANGE, this.object);
+		}
 	}
 	
 	public function removeObject(object:ValEditorObject):Void
@@ -158,7 +225,7 @@ class Selection extends EventDispatcher
 		{
 			this._templateGroup.deleteTemplates();
 		}
-		else
+		else if (this._objectGroup.numObjects != 0)
 		{
 			this._objectGroup.deleteObjects();
 		}
